move sciarg-specific test code from common.py to test_sciarg.py
